+++ foundation/rti-base/src/main/c++/ObjectRoot.cpp
@@ -804,12 +804,12 @@ void ObjectRoot::fromJson(const std::string &jsonString) {
                 break;
             case TypeMedley::FLOAT:
                 (*classAndPropertyNameValueSPMap[classAndPropertyName]).setValue(
-                  static_cast<long>(propertyJSONObject[memberName].asFloat())
+                  propertyJSONObject[memberName].asFloat()
                 );
                 break;
             case TypeMedley::DOUBLE:
                 (*classAndPropertyNameValueSPMap[classAndPropertyName]).setValue(
-                  static_cast<long>(propertyJSONObject[memberName].asDouble())
+                  propertyJSONObject[memberName].asDouble()
                 );
                 break;
             case TypeMedley::STRING:
